Update dynwave.c
Rolls back to v5.1.013 how the inflow/outflow to nodes between a conduit are adjusted for evaporation and seepage losses. It fixes a large WQ mass balance error in v5.1.014/15 as compared to 5.1.013. It also shows the flow loss directly in the conduit where it occurs rather than as just a reduction in the inflow to the conduit's downstream node.